+++ packages/react-querybuilder/src/utils/transformQuery.ts
@@ -5,15 +5,18 @@ import type {
   RuleGroupTypeIC,
   RuleType,
 } from '../types/index.noReact';
+import { isRuleGroup, isRuleGroupType } from './isRuleGroup';
 
 const remapProperties = (
   obj: Record<string, any>,
-  propertyMap: Record<string, string>,
+  propertyMap: Record<string, string | false>,
   deleteRemappedProperties: boolean
 ) =>
   produce(obj, draft => {
     for (const [k, v] of Object.entries(propertyMap)) {
-      if (!!v && k !== v && Object.hasOwn(draft, k)) {
+      if (v === false) {
+        delete draft[k];
+      } else if (!!v && k !== v && Object.hasOwn(draft, k)) {
         draft[v] = draft[k];
         if (deleteRemappedProperties) {
           delete draft[k];
@@ -22,6 +25,9 @@ const remapProperties = (
     }
   });
 
+/**
+ * Options object for {@link transformQuery}.
+ */
 export interface TransformQueryOptions<RG extends RuleGroupTypeAny = RuleGroupType> {
   /**
    * When a rule is encountered in the hierarchy, it will be replaced
@@ -45,18 +51,22 @@ export interface TransformQueryOptions<RG extends RuleGroupTypeAny = RuleGroupTy
    * the new _and_ the original properties, set `deleteRemappedProperties`
    * to `false`.
    *
+   * If a key has a value of `false`, the corresponding property will be removed
+   * without being copied to a new property name. (Warning: `{ rules: false }`
+   * will prevent recursion and only return the processed root group.)
+   *
    * @defaultValue `{}`
    *
    * @example
    * ```
    *   transformQuery(
-   *     { combinator: 'and', rules: [] },
-   *     { propertyMap: { combinator: 'AndOr' } }
+   *     { combinator: 'and', not: true, rules: [] },
+   *     { propertyMap: { combinator: 'AndOr', not: false } }
    *   )
    *   // Returns: { AndOr: 'and', rules: [] }
    * ```
    */
-  propertyMap?: Record<string, string>;
+  propertyMap?: Record<string, string | false>;
   /**
    * Any combinator values (including independent combinators) will be translated
    * from the key in this object to the value.
@@ -93,6 +103,13 @@ export interface TransformQueryOptions<RG extends RuleGroupTypeAny = RuleGroupTy
    */
   operatorMap?: Record<string, string>;
   /**
+   * Prevents the `path` property (see {@link Path}) from being added to each
+   * rule and group in the hierarchy.
+   *
+   * @defaultValue `false`
+   */
+  omitPath?: boolean;
+  /**
    * Original properties remapped according to the `propertyMap` option will be removed.
    *
    * @defaultValue `true`
@@ -110,18 +127,20 @@ export interface TransformQueryOptions<RG extends RuleGroupTypeAny = RuleGroupTy
 }
 
 /**
- * Recursively process a query heirarchy with this versatile utility function.
+ * Recursively process a query heirarchy using this versatile utility function.
  *
- * Documentation: https://react-querybuilder.js.org/docs/api/misc#transformquery
- *
- * @param query - The query to transform
- * @param options - Options
- * @returns The transformed query
+ * [Documentation](https://react-querybuilder.js.org/docs/utils/misc#transformquery)
  */
 export function transformQuery(
   query: RuleGroupType,
   options?: TransformQueryOptions<RuleGroupType>
 ): any;
+/**
+ * Recursively process a query heirarchy with independent combinators using this
+ * versatile utility function.
+ *
+ * [Documentation](https://react-querybuilder.js.org/docs/utils/misc#transformquery)
+ */
 export function transformQuery(
   query: RuleGroupTypeIC,
   options?: TransformQueryOptions<RuleGroupTypeIC>
@@ -136,15 +155,16 @@ export function transformQuery<RG extends RuleGroupTypeAny>(
     propertyMap = {},
     combinatorMap = {},
     operatorMap = {},
+    omitPath = false,
     deleteRemappedProperties = true,
   } = options;
 
-  const processGroup = (rg: RuleGroupTypeAny & { path: number[] }): any => ({
+  const processGroup = (rg: RuleGroupTypeAny): any => ({
     ...ruleGroupProcessor(
       remapProperties(
         {
           ...rg,
-          ...('combinator' in rg
+          ...(isRuleGroupType(rg)
             ? { combinator: combinatorMap[rg.combinator] ?? rg.combinator }
             : {}),
         },
@@ -152,27 +172,32 @@ export function transformQuery<RG extends RuleGroupTypeAny>(
         deleteRemappedProperties
       ) as RG
     ),
-    rules: rg.rules.map((r: any, idx) => {
-      if (typeof r === 'string') {
-        // independent combinators
-        return combinatorMap[r] ?? r;
-      } else if ('rules' in r) {
-        // sub-groups
-        return processGroup({ ...r, path: [...rg.path, idx] });
-      }
-      // rules
-      return ruleProcessor(
-        remapProperties(
-          {
-            ...{ ...r, path: [...rg.path, idx] },
-            ...('operator' in r ? { operator: operatorMap[r.operator] ?? r.operator } : {}),
-          },
-          propertyMap,
-          deleteRemappedProperties
-        ) as RuleType
-      );
-    }),
+    ...(propertyMap['rules'] === false
+      ? null
+      : {
+          [propertyMap['rules'] ?? 'rules']: rg.rules.map((r: any, idx) => {
+            const pathObject = omitPath ? null : { path: [...(rg.path ?? []), idx] };
+            if (typeof r === 'string') {
+              // independent combinators
+              return combinatorMap[r] ?? r;
+            } else if (isRuleGroup(r)) {
+              // sub-groups
+              return processGroup({ ...r, ...pathObject });
+            }
+            // rules
+            return ruleProcessor(
+              remapProperties(
+                {
+                  ...{ ...r, ...pathObject },
+                  ...('operator' in r ? { operator: operatorMap[r.operator] ?? r.operator } : {}),
+                },
+                propertyMap,
+                deleteRemappedProperties
+              ) as RuleType
+            );
+          }),
+        }),
   });
 
-  return processGroup({ ...query, path: [] });
+  return processGroup({ ...query, ...(omitPath ? null : { path: [] }) });
 }